@@ -1,17 +1,17 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 export const useNavStyles = makeStyles({
-	navMenuLink: {
-		color: 'inherit',
-		textDecoration: 'none'
-	},
-	navAppBar: {
-		width: '100vw',
-		marginBottom: '3vh'
-	},
-	mobileTitle: {
-		textAlign: 'center',
-		fontSize: '1.5rem'
-	},
-	mobileMenuButton: {}
+    navMenuLink: {
+        color: 'inherit',
+        textDecoration: 'none'
+    },
+    navAppBar: {
+        width: '100vw',
+        marginBottom: '3vh'
+    },
+    mobileTitle: {
+        textAlign: 'center',
+        fontSize: '1.5rem'
+    },
+    mobileMenuButton: {}
 });
